Fix syntax with logging
Some invocations of logging used the format:
logging.info.()
which includes and extra "." after info.
The correct logging line is:
logging.info()

Also remove logging from install_requires list. logging is part of
the python standard library in python 3. An error is thrown when
trying to install TRANSACT when logging is in the install_requires
list.